Move ResolveHostBase to utils pkg
